<!-- markdownlint-disable -->

## Purpose

This PR serves as a PoC of how we should use the new semantics introduced in #23620 

With this "dumb PoC", we can already see the decrease in the overhead in the scheduler.

## Test result

### Timing the `Scheduler.schedule()` function:

**Before this PR:**
```plaintext
[scheduler.py:54] Average Scheduler.schedule time over last 60 runs: 2.3171 ms
```

**After this PR:**
```plaintext
[scheduler.py:54] Average Scheduler.schedule time over last 60 runs: 1.9629 ms
```

### Nsys profile screenshot

**Before this PR: 8.38 ms when scheduling the prefill requests**

<img width="1938" height="276" alt="image" src="https://github.com/user-attachments/assets/a0ccedb1-d155-4c60-a24f-43b5425ed09b" />


**After this PR: 2.97 ms when scheduling the prefill requests**

<img width="1992" height="276" alt="image" src="https://github.com/user-attachments/assets/f9e23b97-ccca-4269-be2c-f7368f385b49" />


---
<details>
<summary> Essential Elements of an Effective PR Description Checklist </summary>

- [ ] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [ ] The test plan, such as providing test command.
- [ ] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.
- [ ] (Optional) Release notes update. If your change is user facing, please update the release notes draft in the [Google Doc](https://docs.google.com/document/d/1YyVqrgX4gHTtrstbq8oWUImOyPCKSGnJ7xtTpmXzlRs/edit?tab=t.0).
</details>

**BEFORE SUBMITTING, PLEASE READ <https://docs.vllm.ai/en/latest/contributing>** (anything written below this line will be removed by GitHub Actions)